Finished context menu for drawing plots

+++ libEPLViz/src/mainWidgets/QwtBaseModel.hpp
@@ -71,7 +71,7 @@ class QwtBaseModel : public QObject, public BaseModel {
   void maxValueChanged(int notUsed, int nMax);
 
  public slots:
-  virtual void createPlot(uint8_t nodeID, uint16_t index, uint8_t subIndex);
+  virtual void createPlot(uint8_t nodeID, uint16_t index, uint16_t subIndex);
   virtual void setXMin(uint32_t min);
   virtual void setXMax(uint32_t max);
 };
